+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 -----
+## 2.1.30 (2022-10-15)
+- Bug fix on Mac for rendering video files - again
+
+-----
 ## 2.1.29 (2022-10-14)
 - Bug fix on Mac for rendering video files - again
 - Forced render step to 1